UPGRADE POLYFILL
I replace https://github.com/Yaffle/EventSource with
https://github.com/EventSource/eventsource because the original polyfill
had a bug related to closing the eventsource connection, causing a
memory leak in the client (browser).

Also, I replaced all the back ticks (`) in app.js because IE chokes on
them. I had hoped this version of the polyfill would support IE 11, but
it doesn't seem to so the only MS browser support currently is Edge.